chore: clear lint warnings
